feat(dsm-celebrations): add member validation to all endpoints

+++ src/services/dsm/celebrationBoard.services.js
@@ -4,7 +4,8 @@ const { dashboardPrisma } = require('../../prismaClient');
 const selectOnlyValidReactionFields = {
   select: {
     reactionId: true,
-    celebrationId: true
+    celebrationId: true,
+    memberId: true
   }
 };
 
@@ -21,7 +22,7 @@ const selectOnlyValidCelebrationBoardFields = {
     reaction: {
       take: 3,
       select: {
-        userId: true,
+        memberId: true,
       }
     },
     // reaction: {
@@ -32,7 +33,9 @@ const selectOnlyValidCelebrationBoardFields = {
     //     reactionId: true
     //   }
     // },
-    createdAt: true
+    createdAt: true,
+    projectId: true,
+    memberId: true,
   }
 };
 
@@ -46,8 +49,11 @@ const filterByAnonymous = (celebrations) => {
 };
 
 // get list of all celebrations
-const listCelebrations = async () => {
+const listCelebrations = async (projectId) => {
   const celebrations = await dashboardPrisma.Celebration.findMany({
+    where: {
+      projectId
+    },
     orderBy: {
       createdAt: 'desc',
     },
@@ -71,10 +77,11 @@ const getCelebrationById = async (celebrationId) => {
 };
 
 // create a new celebration
-const createCelebration = async (author, content, type, isAnonymous = false, projectId) => {
+const createCelebration = async (author, memberId, content, type, isAnonymous = false, projectId) => {
   const newCelebration = await dashboardPrisma.Celebration.create({
     data: {
       author,
+      memberId,
       isAnonymous,
       content,
       type,
@@ -87,8 +94,20 @@ const createCelebration = async (author, content, type, isAnonymous = false, pro
 };
 
 // update a celebration
-const updateCelebrationById = async (celebrationId, content, type, isAnonymous) => {
-  console.log(celebrationId, content, type, isAnonymous);
+const updateCelebrationById = async (celebrationId, content, type, isAnonymous, memberId) => {
+
+  const celebration = await dashboardPrisma.Celebration.findUnique({
+    where: {
+      celebrationId
+    },
+    select: {
+      memberId: true
+    }
+  });
+
+  if (!celebration) throw new HttpError(404, 'No Record Found');
+  if (celebration.memberId !== memberId) throw new HttpError(403, 'You are not authorized to perform this action');
+
   const updatedCelebration = await dashboardPrisma.Celebration.update({
     where: {
       celebrationId
@@ -96,22 +115,35 @@ const updateCelebrationById = async (celebrationId, content, type, isAnonymous) 
     data: {
       content,
       type,
-      isAnonymous,
+      isAnonymous
     },
     ...selectOnlyValidCelebrationBoardFields
-  }
-  );
-  if (!updatedCelebration) throw new HttpError(404, 'No Record Found');
+  });
+  
   return updatedCelebration;
 };
 
 // delete a celebration
-const deleteCelebrationById = async (celebrationId) => {
+const deleteCelebrationById = async (celebrationId, memberId) => {
+
+  const celebration = await dashboardPrisma.Celebration.findUnique({
+    where: {
+      celebrationId
+    },
+    select: {
+      memberId: true
+    }
+  });
+
+  if (!celebration) throw new HttpError(404, 'No Record Found');
+  if (celebration.memberId !== memberId) throw new HttpError(403, 'You are not authorized to perform this action');
+
   await dashboardPrisma.celebrationReactedUser.deleteMany({
     where: {
       celebrationId
     }
   });
+
   const deletedCelebration = await dashboardPrisma.Celebration.delete({
     where: {
       celebrationId
@@ -123,30 +155,43 @@ const deleteCelebrationById = async (celebrationId) => {
   return deletedCelebration;
 };
 
-const updateReaction = async (celebrationId, userId, isReacted) => {
+const updateReaction = async (celebrationId, memberId, isReacted) => {
+
+  const celebration = await dashboardPrisma.Celebration.findUnique({
+    where: {
+      celebrationId
+    },
+    select: {
+      memberId: true
+    }
+  });
+
+  if (!celebration) throw new HttpError(404, 'No Record Found');
+  if (celebration.memberId === memberId) throw new HttpError(403, 'You are not authorized to perform this action');
+
   const updatedReaction = isReacted ?
     await dashboardPrisma.celebrationReactedUser.create({
       data: {
         celebrationId,
-        userId
+        memberId
       },
       ...selectOnlyValidReactionFields
     }) :
     await dashboardPrisma.celebrationReactedUser.deleteMany({
       where: {
         celebrationId,
-        userId
+        memberId
       },
     });
   if (!isReacted & updatedReaction.count === 0) throw new HttpError(404, 'No Reaction Found');
   return updatedReaction;
 };
 
-const getReaction = async (celebrationId, userId) => {
+const getReaction = async (celebrationId, memberId) => {
   const reaction = await dashboardPrisma.celebrationReactedUser.findMany({
     where: {
       celebrationId,
-      userId
+      memberId
     },
     ...selectOnlyValidReactionFields
   });
